Outsource request auth to oauth library
The oauth library is capable of generating http.Client objects
which automatically wrap requests with the appropriate authen-
tication header. clever-go now relies on this library instead
of setting the header on each request. To cover the case of
APIKey authentication, a BasicAuthTransport struct is added
with methods RoundTrip and Client. When/if APIKey authenticat-
ion is removed from the system, the BasicAuthTransport struct
should be removed.