Expose rate limiting error if passed a 429 (TooManyRequests)
The clever-go client library will now expose the error in the
PagedResults object if a response's status code is 429. More
specifically, when receiving a header of the form that sphinx returns,
the limit and reset data are exposed as well.

Wrote a simple test which responds with a mock 429 and tests that the
appropriate error is exposed.